Refactor and optimize Open Redirect Scanner for improved performance
