--- requirements.txt
@@ -1,29 +1,0 @@
-# Professional Open Redirect Scanner Dependencies
-
-# HTTP and Web Scraping
-aiohttp==3.9.1
-requests==2.31.0
-beautifulsoup4==4.12.2
-lxml==4.9.4
-
-# Browser Automation and Screenshots
-selenium==4.16.0
-webdriver-manager==4.0.1
-
-# JavaScript Analysis
-esprima==4.0.1
-jsbeautifier==1.14.11
-
-# Report Generation
-jinja2==3.1.2
-Pillow==10.1.0
-
-# Utility Libraries
-urllib3==2.1.0
-certifi==2023.11.17
-charset-normalizer==3.3.2
-idna==3.6
-
-# Additional Security Tools
-python-whois==0.8.0
-dnspython==2.4.2
